Fix aio close from non-global namespace
aio [open], as well as similar commands return the name of the
created command.  If this is done in the non-global namespace, the
returned name is implicitly scoped to the current namespace while
the actual command is created in the global namespace. Thus [close]
does not work when invoked in that namespace.

The solution is to return a fully qualified name, such as ::aio.handle3

Note that this may also be a problem for similar command such as
[proc] and [alias] that return command names.

Signed-off-by: Steve Bennett <steveb@workware.net.au>