exec: better handling of pipeline abnormal termination
Consider the command pipeline:

  exec a | b | c

Previously, if any of the subcommands terminated abnormally (with a signal),
the stdout of the pipeline would be lost.

Now the output consists of:

1. standard output from the last command in the pipeline
2. standard error from all commands in the pipeline
3. all abnormal error terminations, if any - but suppressed if any standard error output

In addition, $::errorCode previously always contained the termination status of the last
subcommand, even if it succeeded. Now it contains the termination status of the last
subcommand that failed, or "NONE" if all succeeded.

Additionally, the order of $::errorCode was previously wrong, with pid after the signal id
rather than vice versa.

Signed-off-by: Steve Bennett <steveb@workware.net.au>