exec: read output from pipe before waiting for children
It is necessary to read output from the pipe before waiting
for children to exit to avoid the output filling up and blocking.
The error results still needs to be read after the children have
exited, since it is read from a temp file, not a pipe.

Increase the length of the input to test exec-8.1 to catch
the problem.

Signed-off-by: Steve Bennett <steveb@workware.net.au>