aio: ssl: Allow SNI to be specified
For some SSL connections it is necessary to set the Server Name
Indication in the connection in order to receive the correct
certificate. Allow this as part of the client ssl call with:

  $sock ssl -sni $servername

Also for -server mode, allow the certificate and private key to be
stored in a single file and only be specified once.

Signed-off-by: Steve Bennett <steveb@workware.net.au>