core: improvements to garbage collection
Don't look up every object in the command table, since it can be slow
to do that. Only lookup if the object looks like a reference.

Also, script and expression objects can't contain references that
aren't already contained in sub-objects, so there is no need to
scan them for references.

Signed-off-by: Steve Bennett <steveb@workware.net.au>