configure: Add --allextmod option
To build all non-default extensions as modules if prerequisites are met.

Also move the extension help out of the main help since it is quite long
and show it with --extinfo. Now extension help is generated from $extdb
to avoid the help and the module configuration becoming out of sync.

Signed-off-by: Steve Bennett <steveb@workware.net.au>