build: Need sdl needs 'check false'
If no 'check' is provided, we assume that the checks passed.
If pkg-config is specified, but no other check, use 'check false'
to ensure the module is disable if the pkg-config checks don't pass.

Also travis now does 'make all' so that any module extensions are
also built, and add a dependency on libhiredis-dev to test
building the redis extension.

Signed-off-by: Steve Bennett <steveb@workware.net.au>