dict: Fix possible duplicate entries when setting
Due to the way hash collisions are managed it is possible
to have a sequence where an entry is removed and then another
entry is replaced, however the replacement adds an additional
entry instead of updating the existing entry.

Can be reproduced like this as there is a hash collision between
these two keys:

	dict set d 0,13 X
	dict set d 8,4 Y
	dict unset d 0,13
	dict set d 8,4 Z

Should result in one entry in the dictionary, but instead ends with two.

Signed-off-by: Steve Bennett <steveb@workware.net.au>