file: delete: Fix --force and --
The command line processing was wrong such that 'file delete -force abc'
would also try to delete files named 'delete' and '-force' in the current directory.
If the current directory was writable, this would succeed, but it not writable
this would fail and so the desired target was not deleted.

Signed-off-by: Steve Bennett <steveb@workware.net.au>