aio: gets: Improve behaviour for non-blocking streams
Previously calling gets on a non-blocking stream could easily
result in a partial line. Now if a partial line is read, return zero/empty
to indicate that nothing is available while storing the partial line.
The next call to gets (typically within a readable script) will continue
appending to the previous partial line until a complete line can be returned.

Signed-off-by: Steve Bennett <steveb@workware.net.au>