jim: fix sometimes overly zealous garbage collection
When checking for "weak" references where the only reference is in the
command hash table, need to check the reference count of the key in the hash
table (this is the command name), not the reference count of the object
we are using to look up the command. Without this it is possible that a reference
(typically a lambda) will be collected even though there is still a reference to it.

Signed-off-by: Steve Bennett <steveb@workware.net.au>