oo: better object construction
Now a default constructor is created, as an alias for defaultconstrutor.
The constructor is passed the arguments to new and by default
this accepts a dictionary that is checked for valid instance variables and sets them.

However the constructor can be replaced by one that takes arbitrary arguments.
Thus we can how have:

  a new -optiona -optionb

And the constructor is invoked with arguments '-optiona -optionab'.

This makes object initialisation more flexible.

** Note: This is an incompatible change if you have classes with a constructor
and you create object instances with new <dict>.

Signed-off-by: Steve Bennett <steveb@workware.net.au>