examples.api: drop last JIM_EMBEDDED definitions
The macro JIM_EMBEDDED was required to be defined before including
jim.h in applications that embed jimtcl.
This requirement has been dropped in 2010 with commit 2d8564100c86
("Documentation updates") but it's use has remained for longer and
it even re-appeared later in the examples.api.

Drop last instances of JIM_EMBEDDED.

Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>