core: avoid rare crash on shutdown
If deferred commands run on shutdown as stack frames
are unwound, it is possible that a command
is deleted but a later command still has a reference to it.
So instead of trying to optimise this case by freeing immediately,
cache deleted commands and let the be freed once all stack frames
are destroyed.

Signed-off-by: Steve Bennett <steveb@workware.net.au>