core: improve eval frame handling
Now callers to JimInvokeCommand() are expected to push and eval frame.
Then we no longer need to carry currentScriptObj, argc, argv in the interp
since these are in the current eval frame.

Note that this change simply renames some unused fields in Jim_Interp for ABI
compatibility, but these will be removed in due course.

Signed-off-by: Steve Bennett <steveb@workware.net.au>