core: Display errors in a more "pythonesque" way
A typical error message now looks like this:

    t4.tcl:2: Error: syntax error in expression: "blah"
    Traceback (most recent call last):
      File "t4.tcl", line 14
        c 1 2 3
      File "t4.tcl", line 10, in c
        b a c
      File "t4.tcl", line 6, in b
        a A14
      File "t4.tcl", line 2, in a
        expr blah

This is produced by stackdump (that can be replaced), called by errorInfo.

Note that now stacktraces (stacktrace, info stacktrace, $opts(-errorinfo)) include
the running command at each level in addition to proc, file, line. In order for
scripts to detect this new format, a new entry tcl_platform entry has been added:
tcl_platform(stackFormat) = 4 (to signify 4 elements per frame)

In addition, instead of building the error stack frame as the stack
is unwound in response to an error, instead the entire current stack trace
is captured by stacktrace. This means that the trace extends beyond the try/catch
right back to the initial interpreter command.

The 'stacktrace' command is now implemented in C based on the same
code that generates the error stacktrace.

Signed-off-by: Steve Bennett <steveb@workware.net.au>