Fix abort on boolean negation/plus unary operators
Currently jimtcl will abort if there is any unary operator besides ! in front of a boolean.
It is possible to construct an expression with - or + before a boolean literal (e.g. expr {-true}) that currently aborts, but should return an error.
I implemented this error handling with tcl's error message for this situation.