core: Jim_SetVariable() now frees value on error
If the value has a zero reference count, free it in the error path.
Otherwise callers can't do: Jim_SetVariable(..., Jim_New...())
in case the object is leaked in the error path.

Fix various callers that were previously freeing the object in the
error path, and add a test to loop.test to show that this was not.

Signed-off-by: Steve Bennett <steveb@workware.net.au>